+++ Examples/Resources/AADConditionalAccessPolicy/2-Update.ps1
@@ -68,6 +68,7 @@ Configuration Example
             SignInFrequencyType                      = "hours";
             SignInFrequencyValue                     = 2; # Updated Property
             State                                    = "disabled";
+            TermsOfUse                               = @("Contractor Data Handling Agreement", "Employee Acceptable Use Policy");
             ApplicationId                            = $ApplicationId
             TenantId                                 = $TenantId
             CertificateThumbprint                    = $CertificateThumbprint